voice_retrieval_RAG/requirements.txt
supabase
flashrank
openai
python-dotenv
pyaudio

